pyproject.toml
[tool.poetry]
name = "XaBot"
version = "0.1"
description = "A discord bot multipurpose"
authors = ["FranDev <olmelofran@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.7.3"
python-dotenv = "^0.20.0"
twitchAPI = "^2.5.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
